Bug #680.  Don't accept floats or small ints as var args.
